- Mudada lógica de geração de seres: na classe "Gerador" somente estão os valores que seriam selecionados em uma criação manual de personagem, enquanto todo os cálculos automáticos estão sendo feitos dentro da classe "Valor" - Criada nova classe para abrigar os cáclulos de número: CalculadorNumero - Os seguintes métodos foram mudados de CalculadorSer para CalculadorNumero:  - SomaValorMag  - DivideValorMag  - MultiplicaValorMag  - CalculaPA  - CalculaPorcentagem  - CalculaPorcentagemDetotal - Criados os seguintes métodos e seus testes correspondentes:  - CalculaKarma em CalculadorSer  - CalculaExperiencia em CalculadorSer  - CalculaReacao em CalculadorSer  - GeraReacao em Gerador  - RetornaMaiorValorMag (ValorMag e Lista) em CalculadorNumero  - RetornaMenorValorMag (ValorMag e Lista) em CalculadorNumero  - CriaListaHabilidades em CalculadorSer - Criadas 3 novas propriedades na classe Modificador para controlar a adição ou remoção de Pericias, Habilidades ou Energias baseado em modificadores.  - PericiaAfetada  - HabilidadeAfetada  - EnergiaAfetada

+++ Desire.Core/Ser.cs
@@ -90,9 +90,9 @@ namespace Desire.Core
         public List<Modificador> Fraquezas { get; set; }
 
         //Experiencia
-        public int PontosGraduacao { get; set; }
-        public int PontosEvolucao { get; set; }
-        public int ExperienciaAtual { get; set; }
+        public double PontosGraduacao { get; set; }
+        public double PontosEvolucao { get; set; }
+        public double ExperienciaAtual { get; set; }
 
         //Equipamento
         public List<Item> ItensEquipados { get; set; }
